pyproject.toml
[project]
name = "assistant-agent"
version = "0.1.0"
description = "一个基于 Deepseek V3 & R1 构建的 Agent 对话式应用服务"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14,<3.15"  
dependencies = [
    "uvicorn (>=0.38.0,<0.39.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi (>=0.123.5,<0.124.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "datetime (>=6.0,<7.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "faiss-cpu (>=1.13.1,<2.0.0)",
    "docx (>=0.2.4,<0.3.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "isort (>=7.0.0,<8.0.0)"
]
